@@ -76,18 +76,22 @@
     "rollup-plugin-terser": "^7.0.2",
     "sucrase": "^3.16.0",
     "terser": "^5.6.0",
+    "ts-node": "^9.1.1",
     "tslib": "^2.1.0",
     "typescript": "^4.1.5"
   },
   "devDependencies": {
     "@masquerade-circus/bench-test": "^0.0.2",
     "@release-it/conventional-changelog": "^2.0.0",
+    "@types/node": "^14.14.35",
+    "@typescript-eslint/eslint-plugin": "^4.19.0",
+    "@typescript-eslint/parser": "^4.19.0",
     "compression": "^1.7.4",
     "cross-env": "^7.0.3",
     "cz-conventional-changelog": "3.3.0",
     "dayjs": "^1.9.6",
-    "eslint": "^7.14.0",
-    "eslint-plugin-sonarjs": "^0.5.0",
+    "eslint": "^7.22.0",
+    "eslint-plugin-sonarjs": "^0.6.0",
     "expect": "^26.6.2",
     "faker": "^5.1.0",
     "fastify": "^3.9.1",
