@@ -161,11 +161,11 @@ span.hello{display: inline-block}
 
   it.only("should inline js", async () => {
     v.inline.extensions("ts");
-    await v.inline.ts("./lib/index.ts", { outputOptions: { minify: true } });
-    await v.inline.ts("./lib/index-lite.ts", { outputOptions: { minify: true } });
-    await v.inline.js("./lib/index-old.js", { outputOptions: { minify: true } });
+    // await v.inline.ts("./lib/index.ts.old", { outputOptions: { minify: true } });
+    await v.inline.ts("./lib/index.ts", { outputOptions: { compact: true } });
+    await v.inline.js("./lib/index-old.js", { outputOptions: { compact: true } });
     console.log(v.inline.ts()[0].raw.length);
-    console.log(v.inline.ts()[1].raw.length);
+    // console.log(v.inline.ts()[1].raw.length);
     console.log(v.inline.js()[0].raw.length);
 
     // console.log(v.inline.ts()[1].raw);
